Clean up texture loading a bit
- Add inventory texture
