+++ src/textures.h
@@ -1,0 +1,9 @@
+#ifndef __TEXTURES_H__
+#define __TEXTURES_H__
+namespace konstructs {
+    #define BLOCK_TEXTURES 0
+    #define SKY_TEXTURE 2
+    #define INVENTORY_TEXTURE 4
+    void load_textures();
+};
+#endif
